[ldap] Allow user customization of field mapping
Attributes from a user's LDAP record are used during account-linking to
populate the erchef user record when it is created. Previously, the
mapping between LDAP attributes and chef user attributes were fixed.
Now, they are configurable.

For example, if the user's LDAP record stores their email address in a
field named 'address' instead of 'mail', then you could set the
following in private-chef.rb:

    ldap['email_attribute'] = "address"

Fixes #151
Fixes #800
Fixes #104
Partially addresses #675

Issue #800 was also addressed in #863 which allowed common_name to
service as a fallback for display name. The fallback is still in place
but now any field can be used for the display_name.

Issue #675 is an issue which our unicode handling. The unicode handling
is still broken; however, this would allow users to use a different
field that might not contain multi-byte characters.

Signed-off-by: Steven Danna <steve@chef.io>