[ldap] Don't mangle multi-byte characters
Previously, if a user's ldap data contained multi-byte characters, we
would mangle it.

For example, if a user's displayname was "张伟" it would get returned as
"å¼ ä¼". This was caused by a call to characters_to_binary.

The UTF-8 representation of "张伟" is

    229,188,160,228,188,159

2 characters, each 3 bytes. unicode:characters_to_binary would transform
this into:

    <<195,165,194,188,194,160,195,164,194,188,194,159>>

It does this because characters_to_binary assumes that the input is a
list of unicode code points. In our example, it would expect

    [24352,20255]

and would then return

    <<229,188,160,228,188,159>>

Since we are sending it raw UTF-8 data rather than code points, our data
gets mangled.

Since we are already getting UTF-8 encoded data from eldap, it appears
that we can just forward it on without modification and it will display
correctly.

Fixes #675

Signed-off-by: Steven Danna <steve@chef.io>