Clarify secrets validation logic
We were finding this code confusing to think through so we've changed
it to more explicitly spell out what we think are the valid states and
only raise the error if we didn't see one of those states.

NOTE: During testing we found that the current and new code doesn't
catch the case of the secrets file being deleted between runs. Per a
code comment, this is difficult to test since the check would need to
be able to differentiate between a number of different
scenarios (first run, first run w external setups, etc)

Signed-off-by: Steven Danna <steve@chef.io>